implemented the different function for add, sub, div, mult

+++ fns_and_dsa/arithmetic_operations.py
@@ -1,4 +1,4 @@
-def perform_operation (int: num1, nint: num2, str: operation):
+def perform_operation (int: num1, int: num2, str: operation):
     if operation == 'add':
         return num1 + num2
     elif operation == 'subtract':
